Playing music continuously is it's own tokio task.
Does not quite work yet, as a serenity Context is needed to get the
working instance of Songbird. We don't have access to that context
outside of callback events.

So, on join command, spawn a task and pass a clone of context into the
thread. Then communicate with Server task with channels. Only problem is
how to get the senders and recievers where they need to be...