Recieve data to server and decipher to Vec<u8>. Playback unfinished, not working yet.
Tried using a method for playback that does not work. Although I already
have an idea of what I need to do to get it working, so this is just a
history commit, so I can refer back should I need to.

The idea here was to spawn a thread in the "join" callback, terminate it
in the leave callback. But I could not get the reciever that was
supposed to recieve audio data from the server to the thread properly,
so this did not work. Also overly complicated.