Music now gets to Songbird instance and plays! However...
...the recieved samples are all played at super high speeds, causing
them to overlap. Usually "Blazingly Fast!" is what we need, but not in
this case. Need to rate limit the playback.

The way this works: Instead of a new thread each time join is called,
leverage the amazingness of async and just make one thread that is
instantiated with the Arc<Songbird> ref. Easy to pass over the reciever
and Songbird from "bot_init()" without weird shit happening like last
time. Now it just works!