Recieve audio data in packets of multiple audio frames
Deconstruct frames by length given, and send to `play_sound`. The header
for `Dca` container seems to not be working, but that's okay.

For some reason, because of this, the audio plays at max speed. Need to
fix that...

Took sooooooo long to do this. Very painful. Tried many things, this
finally worked.