Plays the data received over TCP
First reads the guild id from the TcpStream, then moves on to reading
the audio data. After receiving each audio frame, sends back an
acknowledgement to the client, which upon receiving it, sends the next
frame.

Each client is split into its own green thread, and each thread now
calls `bot::play_music(...)` independently, allowing for multiple
clients to be connected simultaneously. To supply the `Songbird`
instance to each of these threads, use a combination of `lazy_static!`
and `OnceCell` to set the variable once in `bot_init` and then access it
everywhere else.

P.S: Just realized while writing this, idk if lazy_static is strictly
necessary, so I'll try to remove it and see if it still works. Should
work, if I understand it properly.