Audio playback updated, and Packets follow DCA1.
Previously the Container and Codec could be specified, but now they are
infered with Symphonia's probe and hint system. This means switching
over to the DCA1 spec with the downside of needing to attach headers and
json metadata to each packet. The upside is that multiple frames can be
sent bundled in a packet now.

Playback works now, but the problem is there is a noticeable moment of
silence between each and every packer. Tested and found that the delay
is whatever goes on in "play_input", not any of the recieve code.