Trying out streaming from custom symphonia source
Create a custom ReadOnlySource impl and try to read the frames in
dynamically using channels to directly send it in. No overhead of
spinning up a new reader every 300ms frame bunch. Implemented the impl
Read part for the new type.

Also, of course, added symphonia to the project explicitly.