Experimental Streaming solution to stuttering audio.
Instead of running songbird's `play_input` for every single packet
recieved as the data is streamed into the server, instead create an
OpusStream struct with custom `Read` implementation that will allow us
to dynamically pass in audio frames as they come in.

Works for a bit then errors. Need to fix.